refactor(database): fix logic
